move common graphics stuff to its own module
